+++ recipes/base_install.rb
@@ -32,10 +32,10 @@ when 'rhel', 'amazon'
     end
   end
   if node['platform'] == 'centos' && node['platform_version'].to_i == 8
-    # Enable PowerTools Repo so *-devel packages can be installed with DNF
-    # Enable EPEL repos
-    execute 'dnf enable powertools and EPEL repos' do
-      command "dnf config-manager --set-enabled PowerTools && dnf install -y epel-release"
+    # Enable powertools repo so *-devel packages can be installed with DNF
+    powertools_repo = find_rhel_minor_version <= '2' ? "PowerTools" : "powertools"
+    execute 'dnf enable powertools' do
+      command "dnf config-manager --set-enabled #{powertools_repo}"
     end
   end
 
@@ -123,15 +123,8 @@ cookbook_file 'AWS-ParallelCluster-License-README.txt' do
   mode '0644'
 end
 
-if node['platform'] == 'ubuntu' && node['platform_version'].to_f >= 16.04
-  # FIXME: https://github.com/atomic-penguin/cookbook-nfs/issues/93
-  include_recipe "nfs::server"
-end
-if node['platform'] == 'centos' && node['platform_version'].to_i == 8
-  # Workaround for issue: https://github.com/atomic-penguin/cookbook-nfs/issues/116
-  node.force_override['nfs']['service']['idmap'] = 'nfs-idmapd'
-end
-include_recipe "nfs::server4"
+# Install NFS packages
+include_recipe "nfs::server"
 
 # Put configure-pat.sh onto the host
 cookbook_file 'configure-pat.sh' do
@@ -230,3 +223,6 @@ include_recipe "aws-parallelcluster::cloudwatch_agent_install"
 
 # Install Amazon Time Sync
 include_recipe "aws-parallelcluster::chrony_install"
+
+# Install ARM Performance Library
+include_recipe "aws-parallelcluster::arm_pl_install"