+++ recipes/dcv_install.rb
@@ -230,3 +230,12 @@ if node['conditions']['dcv_supported']
     end
   end
 end
+
+# Switch runlevel to multi-user.target for official ami
+if node['cfncluster']['is_official_ami_build']
+  if node['init_package'] == 'systemd'
+    execute "set default systemd runlevel to multi-user.target" do
+      command "systemctl set-default multi-user.target"
+    end
+  end
+end