+++ recipes/efa_install.rb
@@ -42,8 +42,8 @@ when 'debian'
 end
 
 installer_options = "-y"
-# efa-kmod currently unavailable for ARM instances
-installer_options += " -k" if arm_instance?
+# skip efa-kmod installation on not supported platforms
+installer_options += " -k" unless node['conditions']['efa_supported']
 # enable gpudirect support
 installer_options += " -g" if efa_gdr_enabled?
 
@@ -56,4 +56,16 @@ bash "install efa" do
     ./efa_installer.sh #{installer_options}
   EFAINSTALL
   not_if { ::Dir.exist?('/opt/amazon/efa') && !efa_gdr_enabled?}
+end
+
+# EFA installer v1.11.0 removes libibverbs-core, which contains hwloc-devel during install
+# hwloc-devel is needed to compile SGE
+# Reinstall hwloc-devel by itself explicitly
+package "install hwloc-devel explicitly" do
+  package_name value_for_platform(
+    'default' => "hwloc-devel",
+    'ubuntu' => { 'default' => "libhwloc-dev" }
+  )
+  retries 3
+  retry_delay 5
 end